requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch>=2.6.0
torchvision>=0.17.0
pillow==10.0.1
python-multipart==0.0.6
numpy>=1.24.0
